requirements.txt
pillow
rembg
colorama